Added export-trac.pl and create-issues.pl
The first reads the Trac sqlite file and produced git-schema'd JSON
objects suitable for import.

The second reads the JSON and performs the Github API requests to
upload the data.  It assumes that the issue list was either empty,
or partially imported form this JSON spec.  (i.e. there must never
have been any issues in the repository before the first run of this
script)